+++ astro.config.mjs
@@ -11,7 +11,7 @@ export default defineConfig({
   integrations: [mdx(), sitemap(), tailwind()],
   output: "hybrid",
   adapter: cloudflare({
-    mode: 'directory',
+    mode: 'advanced',
     routes: {
       include: [
         '/ask'
