+++ package.json
@@ -1,12 +1,12 @@
 {
-  "name": "teal-tower",
+  "name": "iainkirkpatrick.com",
   "type": "module",
   "version": "0.0.1",
   "scripts": {
     "dev": "astro dev --port 3000",
     "functions": "bunx wrangler dev ./_worker.ts --remote",
     "start": "astro dev",
-    "build": "astro check && astro build",
+    "build": "astro check && astro build && bun run build-worker",
     "build-worker": "bun build ./_worker.ts --outdir ./dist --external @cloudflare/ai",
     "preview": "astro preview",
     "astro": "astro"
